% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{antiMerge}
\alias{antiMerge}
\title{Anti join two data.frames}
\usage{
antiMerge(x, y, by)
}
\arguments{
\item{x}{a data.frame}

\item{y}{a data.frame}

\item{by}{a character vector of column names to join by. Defaults to all common columns.}
}
\value{
a data frame with only the rows that are in the first data frame but not in the second data frame, given a set of keys
}
\description{
Returns the rows in \code{x} that are not present in \code{y}
}
\examples{
x <- data.frame(a = c(1, 2, 3), b = c(4, 5, 6))
y <- data.frame(a = c(1, 2, 3), b = c(4, 5, 7))
antiMerge(x, y)
antiMerge(x, y, by = "a")
antiMerge(x, y, by = c("a", "b"))
antiMerge(x, y, by = "b")

}
