% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{callerCall}
\alias{callerCall}
\title{Return the call of the calling function}
\usage{
callerCall(getName = FALSE)
}
\arguments{
\item{getName}{If \code{TRUE}, return the name of the function}
}
\value{
The call of the calling function
}
\description{
Return the call of the calling function
}
