% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{callerEnvironment}
\alias{callerEnvironment}
\title{Return the environment of the calling function}
\usage{
callerEnvironment()
}
\value{
The environment of the calling function
}
\description{
Return the environment of the calling function
}
