% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{callerFrameNumber}
\alias{callerFrameNumber}
\title{Return the frame number of the calling function}
\usage{
callerFrameNumber()
}
\value{
The frame number of the calling function
}
\description{
Return the frame number of the calling function
}
