% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{check_columns}
\alias{check_columns}
\title{Check columns of a data.frame or matrix for strange values}
\usage{
check_columns(df)
}
\arguments{
\item{df}{A data.frame or a matrix object}
}
\description{
Goes through column by column to check for NA's, Dates out of
range, empty strings etc.
}
\examples{
df <- data.frame(a = c(NA, 1, 2), b = c("", "A", NA),
                 stringsAsFactors = FALSE)
check_columns(df)

}
