% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{check_writeable}
\alias{check_writeable}
\alias{file.writeable}
\title{Check if writable}
\usage{
file.writeable(s)
}
\arguments{
\item{s}{CharacterVector path}
}
\description{
Return \code{TRUE} if either a directory or file is writeable.
}
\section{Functions}{
\itemize{
\item \code{file.writeable()}: Check file

}}
\examples{
\dontrun{
file.writeable("example.txt")
}

}
