% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{column_types}
\alias{column_types}
\title{Column classes in a data.frame}
\usage{
column_types(df)
}
\arguments{
\item{df}{A data.frame}
}
\value{
A stacked matrix with two columns
}
\description{
Determine all column types of data.frame
}
\details{
This function takes a data.frame as its single argument.
It then queries all columns for their class(es) and returns a
stacked matrix where the first column indicate class and the second
column column name (taken from df). If a column is represented twice,
it means that the column has more than one class.
}
\examples{
time <- as.POSIXct(Sys.time(),
 tz = Sys.timezone(location = TRUE),
 origin = "1970-01-01")
dd <- data.frame("A" = c(NA, 1), "B" = c(NA,NA), "C" = c(1,3), "time" = time, "F" = factor(c(F)))
column_types(dd)

}
