% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{comma_swap}
\alias{comma_swap}
\title{Swap single comma deliminated strings}
\usage{
comma_swap(v)
}
\arguments{
\item{v}{\code{CharacterVector}}
}
\value{
\code{CharacterVector}
}
\description{
\code{comma_swap} is meant to transform comma deliminated country
names like "Korea, South" to their canonical name, "South Korea".
}
\section{Warning}{
 \code{comma_swap} assumes only a single comma in
each string and will issue an error if there aren't exactly two
elements after splitting on ','.
}

\examples{
comma_swap(c("Korea, South", "Korea, North"))

}
