% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{compare_column_types}
\alias{compare_column_types}
\title{Compare column types across multiple data frames}
\usage{
compare_column_types(...)
}
\arguments{
\item{...}{one or more data frames}
}
\value{
a data frame listing the common columns and their classes in the
input data frames. If there are type clashes, the function returns an
error and a data frame indicating the columns with type clashes.
}
\description{
This function takes one or more data frames as input and compares their
column types. It returns a data frame that lists the columns that are
common to all input data frames, along with the class of each column in
each data frame. If the class of a column is different across the input
data frames, the function returns an error and a data frame indicating
the columns with type clashes.
}
