% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%^\%}
\alias{\%^\%}
\title{concatenate operator}
\usage{
lhs \%^\% rhs
}
\arguments{
\item{lhs}{A character vector.}

\item{rhs}{A character vector.}
}
\description{
Concatenates two strings together using the \code{paste0}
function. \emph{No attempt is made to collapse either vector
argument}.
}
\section{Warning}{
 Because the concatenate operator makes no attempt
to unwind potential recursive calls, there are much more
efficient ways of concatenating large numbers of character
vectors together.
}

\examples{
"Hello " \\%^\\% "World!"
c("a", "b") \\%^\\% c("c", "d")
"foo" \\%^\\% c("a", "b")

}
