% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{df_compare}
\alias{df_compare}
\title{Compare two data.frames rowwise based on anti_joins}
\usage{
df_compare(dfNew, dfOld, keep_cols = FALSE, round_numeric = TRUE)
}
\arguments{
\item{keep_cols}{Logical for keeping all columns from both. Defaults to FALSE.}

\item{round_numeric}{Logical for rounding all numeric columns to three digits prior to joining. If TRUE, POSIXct are rounded to seconds. Defaults to TRUE.}

\item{dfnew}{The new data.frame.}

\item{dfold}{The old data.frame.}
}
\value{
A new data.frame with bi-directional rowwise differences between
\code{dfold} and \code{dfnew}.
}
\description{
Returns the rowwise and bi-directional differences between
\code{dfnew} and \code{dfold} as a new data.frame.
}
\details{
This function takes two data.frames as input and compares them
rowwise using \code{dplyr::anti_join} in both directions (\code{dfnew != dfold} and
\code{dfold != dfnew}). All columns are used for the join, which is not trivial
for input data.frames with columns of class double. As a partial fix, we round
all numeric columns to the third decimal by default prior to joining. If rounding is true,
columns of class POSIXct are rounded to seconds. This is done using \code{vbase::round_df}. To turn this feature off,
set \code{round_numeric = FALSE}. This disables roundig for both classes.
The function inherits \code{dplyr::anti_join} default treatment for missing values.

The output is a new data.frame with the number of rows equal to the number of
rows identified by the joins as only being in either but not both input data.frames.
The default behavior is to only retrieve the intersection of columns. To instead
get the union of columns, set \code{keep_cols = TRUE}. This returns the additional
column with values equal to NA for the data.frame that did not contain this column.

The column \code{dfdiff} takes on two values, - or +. The + means that \code{dfnew}
contains a row that is not in \code{dfold}. The opposite applies to -.
This functionality is best used when sorting on all columns that are of interest
\strong{before} sorting on \code{dfdiff}.
}
\examples{
df_o <- data.frame(id1 = rep(c(1L, 2L), each = 5),
                   id2 = rep(2006L:2010L, 2),
                   value = rnorm(10))

df_n <- rbind.data.frame(df_o[-1, 1:3],
                      data.frame(id1 = c(1L, 2L),
                      id2 = rep(2011L, 2),
                      value = rnorm(2)))
                      
print(df_compare(df_n, df_o), row.names = FALSE)

}
