% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{duplicates}
\alias{duplicates}
\title{Show duplicates from a data.frame}
\usage{
duplicates(df, cols, keep_all = FALSE, ...)
}
\arguments{
\item{df}{A data.frame.}

\item{cols}{A character vector of column names.}

\item{keep_all}{A logical scalar that decides if all columns should be returned.}

\item{...}{Accepts a named vector of additional columns to include in the output.}
}
\value{
A data.frame with the duplicate rows.
}
\description{
Show the duplicate rows of a data.frame based on specified columns.
}
