% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{error}
\alias{error}
\title{Error message}
\usage{
error(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{log_msg}}}
}
\description{
Convenience wrapper around \code{log_msg} to output formatted error
messages.
}
\examples{
info("This is an error message")

}
