% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{fgrepl}
\alias{fgrepl}
\title{Filter a data frame by a regular expression}
\usage{
fgrepl(df, pattern, x, ...)
}
\arguments{
\item{df}{a data frame}

\item{pattern}{a regular expression}

\item{x}{the name of the column in which to apply the regular expression}
}
\value{
a new data frame with only the rows that match the regular
expression in the specified column
}
\description{
This function filters a data frame by a regular expression applied to
the values in a specified column. It returns a new data frame with only
the rows that match the regular expression.
}
