% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{filterRows}
\alias{filterRows}
\title{Filter rows of a data.frame}
\usage{
filterRows(df, ..., combine = "all", includeNA = FALSE)
}
\arguments{
\item{df}{A data.frame}

\item{...}{One or more expressions}

\item{combine}{How to combine the expressions. Either \code{all} or
\code{any}. Defaults to \code{all}.}

\item{includeNA}{Logical for whether to include NA's in the
expressions. Defaults to \code{FALSE}.}
}
\value{
A data.frame with only the rows that match the expressions
}
\description{
Returns a data.frame with only the rows that match the
expressions in \code{...}. The expressions are evaluated in the
environment of \code{df}.
}
\details{
This function takes a data.frame as its first argument and
one or more expressions as its remaining arguments. The expressions
are evaluated in the environment of the data.frame. The expressions
are evaluated rowwise and the function returns a data.frame with
only the rows that match the expressions.
}
\examples{
df <- data.frame(a = c(NA, 1, 2), b = c("", "A", NA),
                stringsAsFactors = FALSE)
filterRows(df, a > 1, b == "A")

}
