% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{first_row_to_names}
\alias{first_row_to_names}
\title{Use the first row of a data frame as column names}
\usage{
first_row_to_names(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a data frame with the first row as column names and the first row
removed
}
\description{
This function takes a data frame as input and sets the names of its columns
to the values in the first row of the data frame. It then removes the first
row from the data frame and returns the modified data frame.
}
