% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{full_rbind}
\alias{full_rbind}
\alias{full_rbind.matrix}
\title{Rbind uneven matrices}
\usage{
full_rbind(x, y)

\method{full_rbind}{matrix}(x, y)
}
\arguments{
\item{x}{A matrix.}

\item{y}{A matrix.}
}
\description{
Rbinds two matrices of different dimensions by first filling
missing columns.
}
\section{Methods (by class)}{
\itemize{
\item \code{full_rbind(matrix)}: S3 method for matrices

}}
\examples{
x <- matrix(1:9, 3, 3)
y <- matrix(1:3, 3, 1)
full_rbind(x, y)

}
