% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{info}
\alias{info}
\title{Information message}
\usage{
info(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{log_msg}}}
}
\description{
Convenience wrapper around \code{log_msg} to output formatted info
messages.
}
\examples{
info("This is an information message")

}
