% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_vars.R
\name{is_path_mounted}
\alias{is_path_mounted}
\title{Check if a path is mounted}
\usage{
is_path_mounted(PATH = Sys.getenv("MOUNT_PATH"))
}
\arguments{
\item{PATH}{Character string with the path to be checked. Defaults to the "MOUNT_PATH" environment variable.}
}
\value{
A logical value indicating whether the path is mounted or not.
}
\description{
This function checks whether a path is mounted in the file system.
}
