% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{leftMerge}
\alias{leftMerge}
\title{Wrapper around merge with all.x = TRUE and a warning if the x object get additional rows}
\usage{
leftMerge(x, y, ...)
}
\description{
Wrapper around merge with all.x = TRUE and a warning if the x object get additional rows
}
