% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{long_to_wide}
\alias{long_to_wide}
\title{long-to-wide reshaping of a data.frame}
\usage{
long_to_wide(df, id_vars, id_var, value_var)
}
\arguments{
\item{df}{A data.frame.}

\item{id_vars}{A character vector of identifier column names that should not be reshaped.}

\item{id_var}{A String with the name of the current identifier column whose data will be
transformed to column names.}

\item{value_var}{A String with the name of the current value column.}
}
\value{
A data.frame with new columns, one for each unique value in \code{id_var}.
}
\description{
Reshape a data.frame from long format to wide format.
}
\details{
This function first converts df to a data.table and then uses
data.table::dcast with aliased arguments. It reshapes to wide based on three arguments:
\code{id_vars} encode the post-reshaping identifying variables, which will identify each row,
\code{id_var} is the current identifying variable, \code{value_var} should be a character
string with the column that holds the values. A message is printed comparing
the old and new dimensions.
}
\examples{
df <- data.frame(
	name = c("v2clacfree", "v2clacfree", "v2blatype", "v2blatype"),
	country_text_id = c("DEU", "DEU", "FRA", "FRA"),
	year = c(2001, 2002, 2001, 2002),
	value = c(1, 1, 2, 2)) 
long_to_wide(df, id_vars = c("country_text_id", "year"),
			id_var = "name", value_var = "value")
}
