% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{mc_assert}
\alias{mc_assert}
\title{Check parallelized results for errors}
\usage{
mc_assert(ll)
}
\arguments{
\item{ll}{List of results}
}
\description{
Given a list of outputs from \code{\link[parallel]{mclapply}} or
\code{\link[parallel]{mcMap}}, asserts whether any errors were
encountered.
}
\details{
Currently, no attempt is made to ascertain the location
within the results list of any errors. Feel free to properly
extend this function.
}
\examples{
\dontrun{
ll <- parallel::mclapply(1:3, sdf)
mc_assert(ll)
}

}
