% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{no_duplicates}
\alias{no_duplicates}
\title{Test for duplicates in a data.frame}
\usage{
no_duplicates(df, cols)
}
\arguments{
\item{df}{A data.frame.}

\item{cols}{A character vector of column names.}
}
\value{
Logical vector of length 1.
}
\description{
Check for duplicate rows of a data.frame based on specified columns.
}
