% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{organiseRows}
\alias{organiseRows}
\title{Order rows of a data.frame}
\usage{
organiseRows(
  df,
  ...,
  naLast = TRUE,
  decreasing = FALSE,
  method = c("auto", "shell", "radix"),
  foreground = FALSE
)
}
\arguments{
\item{df}{A data.frame}

\item{...}{One or more expressions}

\item{naLast}{Logical for whether to put NA's last. Defaults to
\code{TRUE}.}

\item{decreasing}{Logical for whether to sort in decreasing order.
Defaults to \code{FALSE}.}

\item{method}{The sorting method. Either \code{auto}, \code{shell},
or \code{radix}. Defaults to \code{auto}.}

\item{foreground}{Logical for whether to move the ordering variables
to the front of the data.frame. Defaults to \code{FALSE}.}
}
\value{
A data.frame with the rows ordered by the expressions
}
\description{
Returns a data.frame with the rows ordered by the
expressions in \code{...}. The expressions are evaluated in the
environment of \code{df}.
}
\details{
This function takes a data.frame as its first argument and
one or more expressions as its remaining arguments. The expressions
are evaluated in the environment of the data.frame. The expressions
are evaluated rowwise and the function returns a data.frame with
the rows ordered by the expressions.
}
\examples{
df <- data.frame(a = c(NA, 1, 2), b = c("", "A", NA),
               stringsAsFactors = FALSE)
organiseRows(df, a, b)

}
