% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{partial}
\alias{partial}
\title{Partial function application}
\usage{
partial(fn, ...)
}
\arguments{
\item{fn}{A function as symbol or character.}

\item{...}{Arguments to be applied.}
}
\description{
This function creates an enclosure for the call given by \code{fn} and the arguments given by \code{...}. The returned function can be called with additional arguments, which will be appended to the call.
}
\examples{
x <- c(1, 2, NA, 4)
f <- partial(mean, na.rm = TRUE)
ls.str(environment(f))
f(x)

}
