% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{pg_create_table}
\alias{pg_create_table}
\title{Create a database table from a data.frame}
\usage{
pg_create_table(df, name, db, overwrite = FALSE, ...)
}
\description{
Create a database table from a data.frame
}
