% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{pg_drop_table}
\alias{pg_drop_table}
\title{Drop a PostgreSQL table}
\usage{
pg_drop_table(name, db)
}
\arguments{
\item{name}{Character string with the name of the table to drop.}

\item{db}{A PostgreSQL database connection object.}
}
\value{
Returns NULL invisibly.
}
\description{
This function drops a table from a PostgreSQL database.
}
