% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{pg_get_query}
\alias{pg_get_query}
\title{Send SQL query to database and expecting a return}
\usage{
pg_get_query(db, query)
}
\arguments{
\item{db}{Database connection.}

\item{query}{SQL query.}
}
\value{
Returns a data.frame.
}
\description{
If you want to only send a query without expecting a return value, use \code{\link{pg_send_query}}.
}
