% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_admin.R
\name{pg_pass}
\alias{pg_pass}
\title{Use .pgpass file for postgres login}
\usage{
pg_pass(pass_file = file.path(Sys.getenv("HOME"), ".pgpass"), ...)
}
\arguments{
\item{pass_file}{Location of pgpass file. Default: ~/.pgpass.}

\item{...}{Named regular expressions used to filter entries from
~/.pgpass. Names must correspond to one or more of "host",
"port", "dbname", "user", "password".}
}
\description{
Reads login credentials from a pgpass file and returns a list of
named vectors for each entry.
}
\examples{
pg_pass()
pg_pass(host = "example-host", user = "example-user")

}
