% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{pg_replace_table}
\alias{pg_replace_table}
\title{Replace a PostgreSQL table}
\usage{
pg_replace_table(df, name, db)
}
\arguments{
\item{df}{A data frame containing the new data to be inserted into the table.}

\item{name}{Character string with the name of the table to be replaced.}

\item{db}{A PostgreSQL database connection object.}
}
\value{
Returns Null invisibly.
}
\description{
This function replaces a table in a PostgreSQL database with a new data frame.
}
