% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{pg_send_query}
\alias{pg_send_query}
\title{Send SQL query to database without expecting a return value}
\usage{
pg_send_query(db, query)
}
\arguments{
\item{db}{Database connection.}

\item{query}{SQL query.}
}
\value{
Returns NULL invisibly.
}
\description{
Send SQL query to database without expecting a return value
}
