% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{pickAll}
\alias{pickAll}
\title{Ensure to pick all variables}
\usage{
pickAll(...)
}
\arguments{
\item{...}{Variables to pick, either as symbol or character.}
}
\value{
A data frame with the selected variables.
}
\description{
Ensure to pick all selected variables
}
\examples{
df <- data.frame(a = 1:10, b = 1:10, c = 1:10)
selectColumns(df, pickAll(a, b)) # select a and b
selectColumns(df, pickAll(A, b, c)) # fail as A is not in df

}
