% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{pickAny}
\alias{pickAny}
\title{Try to pick any variables}
\usage{
pickAny(...)
}
\arguments{
\item{...}{Variables to pick, either as symbol or character.}
}
\value{
A data frame with the selected variables.
}
\description{
Try to pick any selected variables
}
\examples{
df <- data.frame(a = 1:10, b = 1:10, c = 1:10)
selectColumns(df, pickAny(a, b)) # select a and b
selectColumns(df, pickAny(A, b, c)) # select b and c

}
