% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_file}
\alias{read_file}
\title{Read a given file into \code{R}}
\usage{
read_file(f, fn = NULL, msg = TRUE, ...)
}
\arguments{
\item{f}{File to open.}

\item{fn}{Optional argument to explicitly specify the function R
should use to read the given file.}

\item{\\dots}{Addition arguments passed to the associated function found.}
}
\description{
Convenience wrapper that calls the appropriate function to read a
file into \code{R}.
}
\details{
If no function is provided, \code{read_file} selects a
function from a hardcoded list depending on the suffix of the
provided file. If a function is not found, an error will be
raised.
}
