% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{render_snippet}
\alias{render_snippet}
\title{Render a snippet of moustache templates for each row of a data.frame}
\usage{
render_snippet(df, snippet)
}
\arguments{
\item{df}{A data.frame}

\item{snippet}{A character vector of moustache templates}
}
\value{
A character vector of rendered snippets
}
\description{
This function takes a data.frame and a snippet of moustache templates
and renders the snippet for each row of the data.frame. The column names are matched
for the moustache identifiers. See https://mustache.github.io/ for details
}
