% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{round_df}
\alias{round_df}
\title{Round the numeric columns of a data.frame}
\usage{
round_df(df, n)
}
\arguments{
\item{df}{A data.frame}

\item{n}{Decimal position for rounding}
}
\value{
A data.frame
}
\description{
Returns a data.frame with all numeric colums to the nth digit.
}
\details{
This function rounds all numeric and all POSIXct columns in \code{df}.
For numeric columns, it rounds to the \code{nth} digit; for POSIXct columns, it
rounds to seconds (shaving off trailing seconds since Unix Epoch). Note that it rounds,
and not truncate. All other columns are left as is. The returned object is a data.frame.

It checks for two conditions:
\enumerate{
\item \code{is.numeric(column) == TRUE}, and
\item \code{class(column)[1] == "POSIXct"}
}
}
\examples{
df <- data.frame(
  nmr = as.numeric(log(10)),
  time = as.POSIXct(Sys.time(), tz = Sys.timezone(location = TRUE), origin = "1970-01-01"),
  char = LETTERS[5]
  )

print(df)                      
print(round_df(df = df, n = 3))

}
