% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{selectColumns}
\alias{selectColumns}
\title{Select columns of a data.frame}
\usage{
selectColumns(df, ..., keep_all = FALSE, drop = FALSE)
}
\arguments{
\item{df}{A data.frame}

\item{...}{One or more column names}

\item{keep_all}{Logical for whether to keep all columns. Defaults to
\code{FALSE}.}

\item{drop}{Logical for whether to drop the data.frame if only one
column is selected. Defaults to \code{FALSE}.}
}
\value{
A data.frame with only the columns specified in the
arguments
}
\description{
Returns a data.frame with only the columns specified in
\code{...}. The columns are evaluated in the environment of
\code{df}.
}
\details{
This function takes a data.frame as its first argument and
one or more column names as its remaining arguments. The column names
are evaluated in the environment of the data.frame. The function
returns a data.frame with only the columns specified in the
arguments.
}
\examples{
selectColumns(mtcars, mpg, cyl)
selectColumns(mtcars, mpg, cyl, keep_all = TRUE)
selectColumns(mtcars, mpg, drop = TRUE)
selectColumns(mtcars, pickAny(mpg, cyl))
selectColumns(mtcars, pickAll(mpg, cyl))
selectColumns(mtcars, pickAny(mpg, cyl), keep_all = TRUE)
selectColumns(mtcars, pickAll(mpg, cyl), keep_all = TRUE)
selectColumns(mtcars, pickAny(mpg, cyl), drop = TRUE)
selectColumns(mtcars, pickAll(mpg, cyl), drop = TRUE)

}
