% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session_info}
\alias{session_info}
\title{Print information on session and machine}
\usage{
session_info()
}
\value{
Returns output from \code{sessionInfo()}
}
\description{
This function prints information on the session and machine.
Use capture.output to capture the output and prevent printing to the console.
}
