% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{try_error}
\alias{try_error}
\alias{is.try_error}
\alias{get_errmsg}
\title{\code{try-error}}
\usage{
is.try_error(x)

get_errmsg(e)
}
\arguments{
\item{x}{Object to test}

\item{e}{\code{try-error} object}
}
\description{
Convenience functions for dealing with error object results from a
failed \code{try} expression.
}
\section{Functions}{
\itemize{
\item \code{is.try_error()}: Indicates whether an object is a
\code{try-error}.

\item \code{get_errmsg()}: Returns the error message associated with a
try-error object.

}}
\examples{
x <- try(.unknown_function(), silent = TRUE)
is.try_error(x)

}
