% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{untibble}
\alias{untibble}
\title{Remove tibble classes from a data frame}
\usage{
untibble(tbl)
}
\arguments{
\item{tbl}{a data frame}
}
\value{
a data frame with tibble classes removed
}
\description{
This function takes a data frame as input and removes all tibble-related
classes from it, including "tbl_df" and "tbl". If the input is not a
data frame, the function returns an error.
}
