% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{warn}
\alias{warn}
\title{Warning message}
\usage{
warn(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{log_msg}}}
}
\description{
Convenience wrapper around \code{log_msg} to output formatted warning
messages.
}
\examples{
warn("This is a warning")

}
