% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{wide_to_long}
\alias{wide_to_long}
\title{wide-to-long reshaping of a data.frame}
\usage{
wide_to_long(df, id_vars = NULL)
}
\arguments{
\item{df}{A data.frame.}

\item{id_vars}{A character vector of identifier column names that should not be reshaped. Alias for \code{id.vars}.}
}
\value{
A data.frame with two new columns additional to \code{id_vars},
\code{variable} for former column names, and \code{value} for the corresponding values.
}
\description{
Reshape a data.frame from wide format to long format.
}
\details{
This function first converts df to a data.table and then uses
data.table::melt with aliased arguments. It pivots all
variables that are not specified in \code{id_vars}. Hence, measure.vars
gets all columns that are not specified in id_vars. A message is printed
}
\examples{
df <- data.frame(
	country_text_id = c("DEU", "DEU", "FRA", "FRA"),
	year = c(2001, 2002, 2001, 2002),
	v2clacfree = c(1, 1, 2, 2),
	v2blatype = c(3, 3, 4, 4))
wide_to_long(df, id_vars = c("country_text_id", "year"))
}
